+++ package.json
@@ -10,7 +10,8 @@
     "@libsql/client": "^0.8.1",
     "drizzle-kit": "^0.23.2",
     "drizzle-orm": "^0.32.2",
-    "hono": "^4.5.4"
+    "hono": "^4.5.4",
+    "resend": "^4.0.0"
   },
   "devDependencies": {
     "@cloudflare/workers-types": "^4.20240529.0",
